# examples/old/01-tutorial/13-structs/docs/openapi/BookService.yaml
openapi: 3.1.0
info:
  title: library.BookService API
  version: 1.0.0
paths:
...
  /api/v1/categories/:id:
    get:
      operationId: get_category
      summary: get_category
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
    put:
      operationId: update_category
      summary: update_category
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
    delete:
      operationId: delete_category
      summary: delete_category
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT